+++ skellysnapshot/__init__.py
@@ -1,0 +1,7 @@
+import logging
+
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s || %(levelname)s || %(funcName)s || %(message)s',
+    datefmt='%Y-%m-%d %H:%M:%S'
+)
